+++ packages/docs/.storybook/webpack.config.js
@@ -34,7 +34,17 @@ module.exports = ({ config }) => {
     ]
   });
 
-  config.resolve.extensions.push('.ts', '.tsx');
+  config.module.rules
+    .push({
+      test: /\.scss$/,
+      use: [
+        { loader: "style-loader" },
+        { loader: "css-loader" },
+        { loader: "sass-loader" }
+      ]
+    });
+
+  config.resolve.extensions.push('.ts', '.tsx', 'js', 'jsx');
 
   config.resolve.alias = {
     'react-native': require.resolve('react-native-web'),