+++ packages/styles/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.0.1 (2019-10-15)
+
+**Note:** Version bump only for package @natds/styles
+
+
+
+
+
+## [0.0.5](https://github.com/natura-cosmeticos/natds/compare/@natds/styles@0.0.4...@natds/styles@0.0.5) (2019-10-15)
+
+**Note:** Version bump only for package @natds/styles
+
+
+
+
+
+## 0.0.4 (2019-10-15)
+
+**Note:** Version bump only for package @natds/styles
+
+
+
+
+
 ## [0.0.3](https://github.com/natura-cosmeticos/natds/compare/@natds/styles@0.0.2...@natds/styles@0.0.3) (2019-10-01)
 
 **Note:** Version bump only for package @natds/styles